feat(auth): add 401 handling with refresh token
